+++ src/com/intellij/idea/plugin/hybris/toolwindow/CopyFileToHybrisConsoleDialog.java
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -24,7 +24,6 @@ import org.jetbrains.annotations.Nullable;
 
 import javax.swing.*;
 
-import static com.intellij.openapi.ui.DialogWrapper.IdeModalityType.PROJECT;
 
 public final class CopyFileToHybrisConsoleDialog extends DialogWrapper {
 
@@ -34,7 +33,7 @@ public final class CopyFileToHybrisConsoleDialog extends DialogWrapper {
         @Nullable final Project project,
         @Nullable final String dialogTitle
     ) {
-        super(project, false, PROJECT);
+        super(project, false, IdeModalityType.IDE);
         setTitle(dialogTitle);
         init();
     }
@@ -45,7 +44,7 @@ public final class CopyFileToHybrisConsoleDialog extends DialogWrapper {
     }
 
     public void show(final Runnable runnable) {
-        if(showAndGet()) {
+        if (showAndGet()) {
             runnable.run();
         }
     }